fixes #83 by removing storage and retrieval prevConfig as it did not work
Edits to the config file where not captured as code used previous config file. By removing this step we go back to how the code was and force the user to always specify the config file manually, which ensures that there can be no misunderstanding about what config file was used.